tests: cleanup KataConfig after test
After the test we should clean up to avoid conflict with other tests.
This is not the optimal way since there is too much duplicated code. I
will try to fix this properly with an AfterEach(func()). Just adding
this now, since it is blocking the main commit (the next one).

Signed-off-by: Beraldo Leal <bleal@redhat.com>