replace (more or less) arbitrary waits with two fundamental ones
So far, the installation procedure was requeuing on multiple occassions,
mostly to wait for something and just to stay of the safe side.

Ideally, we should be able to replace all this waiting with a single wait
that actually matters to us - the wait for the MCO to finish updating kata
installation on cluster.  However, as it turns out, we need to add one more
wait for implementation reasons, a wait for the MCO to actually *start*
processing cluster changes before we can wait for the MCO to *finish*
the processing.

That's what this commit does.  We first figure out whether kata
installation on the cluster is about to change.  This is done by seeing
if we changed node-role labeling on a regular cluster and if we've just
created the kata extension MC on a converged cluster.  If so we record
this in newly introduced KataConfig.status.waitingForMcoToStart and wait
for the relevant MCP(s) to actually start updating.  Once they do we wait
for them to settle again.

Most requeue removals in this commit are grounded in the MCO operation and
resource watching mechanics.  In this context, the removal of the wait
after the extension MC creation is notable as its rationale comes from a
different source, namely an assumption that the API Server is sequentially
consistent given that etcd is sequentially consistent(*).  This seems
reasonable and confirmed by testing so far.

(*) https://etcd.io/docs/v3.3/learning/api_guarantees/

Fixes: KATA-1969

Signed-off-by: Pavel Mores <pmores@redhat.com>