Apply peerpod related MachineConfigs
Peer pod enablement requires creating a new kata runtime class
and providing the configuration.toml. In OpenShift we need to create
MachineConfig objects to create these files.
The required MachineConfig yaml files are bundled within the
controller-manager image and the manager process reads the yaml
and creates the required MachineConfig objects. This gives us the
flexibility to keep updating the yaml as needed without hardcoding the
struct in the code

Fixes: #Kata-2130

Signed-off-by: Jens Freimann <jfreimann@redhat.com>
Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>