deploy the same flow in uninstallation as is used by installation
This is the central commit of this series.  It replaces several arbitrary
60 seconds waits and node-counting conditionals which have proved fragile
in practice with a straightened-out flow that's identical to what
installation uses: first, change MCO resources (here by unlabeling nodes
and deleting the extension MC), then wait for the MCO to notice the changes
and start processing them, then wait for the MCO to finish processing the
changes, and finally finish up the OSC part of the operation by deleting
the custom pool if applicable, the Daemonset and the Scc, and then
KataConfig itself.

Signed-off-by: Pavel Mores <pmores@redhat.com>