streamline KataConfig.status updating during uninstallation
'r.Client.Status().Update()' was called right after updateStatus() call,
basically duplicating the same catch-all call performed after
processKataConfigDeleteRequest() return.

The processKataConfigDeleteRequest() call site was modified to ensure that
the status update is always performed, even if a deletion reconciliation
ends in an error.  This seems to make sense, no matter how reconciliation
ends it might well have updated the status, and if not the update is just
a no-op.

Signed-off-by: Pavel Mores <pmores@redhat.com>