Adapt `make test` to PR #300
When it comes to deleting the KataConfig, the controller now monitors
some transitions in the statuses of the MCPs.

Implement these in the test code in order to avoid the controller to
loop on the `Waiting for MCO to start updating` condition until the
test times out.

Fixes https://issues.redhat.com/browse/KATA-2167

Signed-off-by: Greg Kurz <groug@kaod.org>