Revert "Fix errors during envtest teardown"
`make test` is leaking the kube-apiserver and etcd processes.
When unoticed, this causes `make test` to fail with weird
errors like failing to open a file because of rlimit.

Investigation leads to commit 053f8664bf29d which was added
to work around the very same problems at the time according to
Pradipta. Reverting this commit fixes the issue.

It seems that updates to the controller-runtime and deps might
have changed something and we no longer need the workaround.

Let's just revert 053f8664bf29d.

Fixes https://issues.redhat.com/browse/KATA-2168

Signed-off-by: Greg Kurz <groug@kaod.org>