fix KataConfig deletion when there are no kata nodes
This fixes an omission in PR #300 where the special case of KataConfig
deletion while there are no kata nodes on a cluster wasn't handled
properly and uninstallation got stuck.  This happened because the
uninstallation flow assumed incorrectly that it will always be
necessary for the MCO to reconciliate which is not the case.

Installation flow (PR #291) got this right and while the idea of PR #300
was basically to make uninstallation flow analogous to installation, this
aspect was omitted by mistake.

Signed-off-by: Pavel Mores <pmores@redhat.com>