Fix envtest tearing down
PR #307 was too optimistic. The original issue worked around
by 053f8664bf29d still exists actually ; `make test` is now
failing with :

  [FAILED] Unexpected error:
      <errors.aggregate | len:1, cap:1>: [
          <*errors.errorString | 0xc0005ebf80>{
              s: "timeout waiting for process kube-apiserver to stop",
          },
      ]
      timeout waiting for process kube-apiserver to stop
  occurred

This is caused by a missing cancellation of the certwatcher in the
controller-runtime. The fix, as documented in [1], is to create
a cancellable context, pass it to envtest and cancel it just before
tearing down the test.

[1] https://github.com/kubernetes-sigs/kubebuilder/pull/2379

Signed-off-by: Greg Kurz <groug@kaod.org>