Allow to skip `go test`
`make test` has been broken since #300 was merged and fixing it is going to
take time (see #310 for details). It is a problem for local development since
the `test` target is a prequisite for `docker-build` in the makefile.

Add a variable to control whether `go test` should run when the `test`
target is invoked.

 $ make test SKIP_TESTS=   # Run `go test` when unset

 $ make test SKIP_TESTS=1  # Don't run `go test` when set

Default is unset.

Signed-off-by: Greg Kurz <groug@kaod.org>