add new Done Nodes status updating
If the MCO considers a Node Done, we check if its current and target
MachineConfigs are the same.

If they are then the Node is actually done and we append it either
to the installation completed list or uninstallation completed list based
on whether installation or uninstallation is in progress and whether
the Node is a member of "worker" or "kata-oc".

If they aren't it means that the Node is waiting for its MC to be changed
as part of this ongoing installation or uninstallation.  We don't append
such Nodes to any list since there doesn't seem to be a good place for
them in the current KataConfig.status structure.

Signed-off-by: Pavel Mores <pmores@redhat.com>
(cherry picked from commit 140307e484a64f703979d63006d5b7afe196b2f8)
Signed-off-by: Greg Kurz <groug@kaod.org>