add new Working/Degraded Nodes status updating
Working and Degraded Nodes are handled structurally similarly so they are
handled by the same function.

Working Nodes are put on an InProgress list, Degraded Nodes are put on
a Failed list.

If uninstallation is in progress Nodes go to uninstallation lists.
If installation is in progress they go to installation lists on converged
clusters or if they are labeled for kata, otherwise they go to
uninstallation lists (this handles the case when kata is removed from
a Node as part of kata installation modification on a cluster).

Signed-off-by: Pavel Mores <pmores@redhat.com>