add new status update main function
This puts together functionality added by the previous commits.

The approach to KataConfig.status updating introduced here fundamentally
differs from the existing updateStatus() function.

The existing function only handled kata installation to and uninstallation
from a cluster with no support for modifications (adding/removing
kata-enabled Nodes) in the meantime.  Thus it made sense to split status
updating at the top level into two largely independent branches, one for
installation and the other for uninstallation, each of which then iterated
over Nodes and updated status based on the Nodes' MCO 'state' and other
information.

This model however breaks when kata installation modifications are
introduced since they mean that kata may actually be uninstalled from Nodes
as part of an operation that is nominally an installation.

The approach introduced here kind of turns the existing one inside out.
It iterates over *Nodes* at the top level and only then it figures out
for each Node individually whether kata is being installed on it,
uninstalled from it, or if the Node is left alone by the ongoing operation.

Signed-off-by: Pavel Mores <pmores@redhat.com>