turn on the new status updating, remove the legacy one
Any errors from the new status updating are just logged, status updating
thus cannot mess up installation/uninstallation flow anymore.  This is
appropriate as status updating now does just that, update status, so any
failures should not divert the main flow.

This also removes the last opportunity for races like the recent
uninstallation wedging due to updateStatus() failing to retrieve the
"kata-oc" MCP which has already been deleted the uninstallation flow,
and thus forcing the uninstallation flow to return early with an error and
never to be able to finish.

Signed-off-by: Pavel Mores <pmores@redhat.com>