final clean-up
With removal of the legacy updateStatus(), updateStatusNew() can be renamed
to the original name.

With removal of the legacy updateStatus() infrastructure getMcp() is left
with no callers and can be removed.

With removal of getMcp() the last remaining caller of
getMcpNameIfMcpExists() is removed so this function can finally be removed
too.  This concludes a refactor started in commit 3adfbd53db.

Signed-off-by: Pavel Mores <pmores@redhat.com>
(cherry picked from commit b8a72355dcfd07124e72c4d6ff72685893926a33)
Signed-off-by: Greg Kurz <groug@kaod.org>