remove legacy status fields usage from uninstallation flow
Handling of uninstallation being blocked by running kata pods turned out
to fit the InProgress Condition neatly by introducing a new
"BlockedByExistingKataPods" reason for _not_ being in-progress.

The rest of the migration is pretty straightforward - setting the legacy
IsInProgress boolean to True corresponds to
setInProgressConditionToUninstalling(), setting it to False corresponds to
resetInProgressCondition().

One place that set the boolean to True was spurious (it was set to True
already) so it's removed without replacement.

Signed-off-by: Pavel Mores <pmores@redhat.com>