remove legacy status fields usage from installation flow
While the idea is simple and the same as with the uninstallation flow
in the previous commit, installation actually did use
InstallationStatus.IsInProgress in its implementation so the replacement
is a bit more tricky.

It's solved by observing that the legacy field was only set if the MCO
was updating "kata-oc" and/or "worker".  We can do the same for a newly
introduced isInstallationInProgress variable whose semantics are basically
the same as InstallationStatus.IsInProgress', just without persisting it
in the KataConfig resource.

Signed-off-by: Pavel Mores <pmores@redhat.com>