stop updating legacy parts of KataConfig.status
The "is in progress" check at the top of updateStatus() is now simplified,
kata-enabled node count can now be assigned directly to
Status.KataNodes.NodeCount.  Updating the legacy structures is just turned
off and proper error handling is added to the putNodeOnStatusList() call
as this function now bears the brunt of updateStatus().

Signed-off-by: Pavel Mores <pmores@redhat.com>