Add feature for creating additional runtimeClasses
This is enabled via AdditionalRuntimeClasses feature gate.
Example configuration is shown below:

apiVersion: v1
kind: ConfigMap
metadata:
  name: osc-feature-gates
  namespace: openshift-sandboxed-containers-operator
data:
  AdditionalRuntimeClasses: "true"

---
apiVersion: v1
kind: ConfigMap
metadata:
  name: additionalruntimeclasses-config
  namespace: openshift-sandboxed-containers-operator
data:
  runtimeClassConfig: "name1:cpuOverHead1:memOverHead1, name2:cpuOverHead2:memOverHead2"
  #runtimeClassConfig: "name1, name2"

Also rearrange the code to have all the runtimeClass handling in a
single code file for easier readability

This feature typically will be used with other features.
For example this feature can be used with the image based deployment feature as well.
The image based deployment might have configuration for multiple
runtimeclasses which is then exposed to users as RuntimeClass objects
created by OSC operator when this feature gate is enabled.

Logs an error if AdditionalRuntimeClasses feature is enabled
and config is not provided

Fixes: #KATA-2903
Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>