Remove duplicate code for reading yamls
We had different methods to read yaml files depending on the
kubernetes object. However this is really not needed as all the methods
simply were reading files with the only difference being the flie
location.
This commit simplifies the code and uses a single readYamlFile method.
The caller should set the full path of the yaml file to be read.
Further any additional requirements to read different yaml files will
no longer need a new method

Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>