Stop using `latest` tag
When an operator version is released, we want it to use images with the
same version exclusively, not `latest` because it could be changed to
point to some other version later on.

Let's do that always in order to avoid subtle bugs when creating custom
catalogs. Current version on devel is 1.6.0.

Signed-off-by: Greg Kurz <groug@kaod.org>