Merge featuregates.go in fg_handler.go
Motivation:
There was significant overlap between the two files.
It was not clear why the two files were split to begin with.
Implementating of a new feature gate would've required adding
new definitions and references to both files. This was
not desirable, as we wish to have the code for each feature as
separated as possible from the rest of the code.

Signed-off-by: Camilla Conte <cconte@redhat.com>