metrics: add initial support for SLI/SLO metrics
Introduces high-level SLI metrics that can help derive SLOs for
kata-remote.

Currently, the osc-monitor image only includes the kata-monitor binary.
The idea is to reuse the same image to also include the metrics-server.

For now, the idea is to collect metrics for kata-remote, but we can
expand later to get more metrics.

Signed-off-by: Beraldo Leal <bleal@redhat.com>