azure: Delete source image on cleanup
OSC creates the podvm image with packer by default and publishes it in a
shared gallery. When the KataConfig CRD is deleted, OSC removes the gallery
but the source image created by packer is left behind. This can generate
uneeded storage costs to the customer.

Fortunately, the path to the source image is available in the output of
`az sig image-version show`. Fix the podvm image deletion flow to also
delete the source image.

Fixes: https://issues.redhat.com/browse/KATA-3462

Signed-off-by: Greg Kurz <groug@kaod.org>