controller: Ensure image creation job runs on an installed node
This potentially limits the issues that may be caused due to
node reboot due to Kata rpm installation
Ideally we need to handle the image creation (and subsequent lifecycle
management) via a proper CRD and state management but we don't do it.
Consider this as an interim fix till we have a proper image lifecycle
management support.

Fixes: #KATA-3573
Signed-off-by: Pradipta Banerjee <pradipta.banerjee@gmail.com>