konflux: avoid building the operator when manager.yaml is updated
We get automated PRs to update the digest of the images in our
ClusterServiceVersion and in the manager.yaml file.
These PRs are generated when one of the referenced images is updated.
Now if modifying those files triggers another build, we can end up
in a loop of builds/nudges.
This commit makes sure that that operator is not rebuilt when such PRs
are merged.

Signed-off-by: Julien Ropé <jrope@redhat.com>