konflux: regular expression fixes for on-cel-expression
All my regexp were wrong. Some worked for the wrong reason, but I was
still triggering the pipelines on wrong file changes on some occasions.

I've used a regexp tester and used the entries of our on-cel-expression
against the actual content of our repository, and came up with the
changes from this commit.
Hopefully, this should be working now.

Signed-off-by: Julien Ropé <jrope@redhat.com>