Adding a test catalog builder
The goal of this special catalog is to provide a very simple catalog
that contains only the bundle for the latest build.
This is not a catalog we will use for release. The goal is to have
it built as a component of the sandboxed-containers application,
nudged by the updates of the bundle.
Hopefully, this should automate the generation of a simple catalog
that our testers can use.

Signed-off-by: Julien Ropé <jrope@redhat.com>