test-fbc: run the render.sh script as part of the build.
We can "nudge" the test-fbc component when the bundle is rebuilt.
But that will just update the bundle reference, not all the related images.

Rather than nuding this component from all images, I'm running the "render.sh"
script when we build this FBC. This way, when the bundle reference is updated,
the catalog should be updated and built with all related images.

Signed-off-by: Julien Ropé <jrope@redhat.com>