tests: add the "make test" check
This commit creates a new integration test that takes the sources of the
operator as they are in the running pull request, and runs "make test" on
them.

This integration test should be set to run only when the operator's sources
were modified - it doesn't make sense to run "make test" on unrelated changes.

Signed-off-by: Julien Ropé <jrope@redhat.com>