Fix kustomization files
The kustomization files allow to customize the operator image and
version in the bundle. They are actionned when running `make bundle`.

We don't rely on that with the konflux builds. It is only useful
when doing custom builds of the operator from a local repo. Let's
bump the version to 1.10 there like everywhere else.

Signed-off-by: Greg Kurz <groug@kaod.org>