bundle: add s390x as supported arch in the labels
When releasing 1.10 using Konflux, we found s390x clusters could not see
our operator in OperatorHub - even though all images were accessible, and
could be installed with CLI.

It seems this is because of the labels we put on the bundle and FBCs.
Fixing it here, will need an additional fix for the FBC part with newly
built bundle image reference.

Signed-off-by: Julien Ropé <jrope@redhat.com>