controllers: create kata-cc-* runtime classes
This adds initial support for Confidential on baremetal. Until now,
Confidential was only supported by peerpods. With this change, we are
introducing the Confidential feature to baremetal as well.

Our consensus is to use the same config option "confidential" to enable
Confidential features in both cases.

We are also reusing the same logic for KataConfigPoolSelector and
checkNodeEligibility, just as we have been doing so far.

Currently, intel and amd labels uses the same string as defined at
scripts/install-helpers/baremetal-coco/nfd/intel-rules.yaml#L26 and
scripts/install-helpers/baremetal-coco/nfd/amd-rules.yaml#L10. Ideally
we should be using a NFD standard string for that.

All the CRI-O and kata-containers configurations will be installed by
default on all nodes , regardless of whether the machine has support.
This will be RPM responsability. The RuntimeClass would be responsible
for filtering machines according to their capabilities.

Related: KATA-3967.

Signed-off-by: Beraldo Leal <bleal@redhat.com>