pipeline definition: adding documentation and template files
This commit provides our pipeline definitions, that will allow us to duplicate
our pipelines in order to support concurrent work on different versions of OSC,
while having CI ready for each branch.

Those files are not part of the build, and are not used by the Konflux pipelines
directly. We need to "apply" them from the CLI to our Konflux instance when we
want to duplicate our pipeline for a building from a new branch.
I am commiting those files in our repo to keep them available, so that we can do
that process again in the future, if need be.

The README.md file provides background information, and the steps we need to follow.

Signed-off-by: Julien Ropé <jrope@redhat.com>