Refactor peer pods disablement code
Consolidate the peer pods disablement code in a separate function.

Ideally we'd want to move all the code to disablePeerPods for
symmetry with enablePeerPods. Unfortunately, this would require
to change the function signature and thus make the patch bigger.
To keep the review simple, rename the existing disablePeerPods so
that the name is available for the new code.

Signed-off-by: Greg Kurz <groug@kaod.org>