image_generator: add function checkAnyKeyPresentWithValue
We already have a helper function that checks that all keys from a given set are present with
a non-nil value. Now add a new helper that checks if any key out of of a given set is present with
a non-nil value. This will be required to support IBM Cloud authentication.

Factor out the map conversion code to a function that can be used by the two check functions.

Signed-off-by: Patrik Fodor <patrik.fodor@ibm.com>