test-fbc: set the catalog's tag to 1.11.0-timestamp
Keeping the tag to 1.10.1 is confusing when bugs are reported from this catalog.
Using 1.11.0 as this is our next release. We'll need to bump it again when it
comes out, to avoid confusion.

This commit does not change the version strings everywhere else, it is just
about the test catalog.

Signed-off-by: Julien Ropé <jrope@redhat.com>