controller: Add overhead annotation using pod mutating webhook
In order for the Kata runtime to correctly deal with hot plugging and
set a VM size that is compatible with the host cgroup size [1], we
need to pass the known overhead (from the RuntimeClass) to the runtime
using an annotation.

This code implements a mutating web hook to modify the pod so that the
annotation is added when the runtime class is Kata.

Signed-off-by: Christophe de Dinechin <dinechin@redhat.com>