Add operator image to manager.yaml
The current operator image is only stored in the CSV within the `bundle`
directory that will be removed from the git repo soon. Record the image
as a related image in manager.yaml and have the tekton pipelines to
monitor it for updates. This will be used to build the bundle with
konflux.

Signed-off-by: Greg Kurz <groug@kaod.org>