fbc: use local build pipeline definition
Using the remote reference to fbc build is causing an issue, with the
Enterprise Contract complaining about obsolete tasks that need to be
updated.
Setting the latest remote task reference didn't help, so I am reverting
it to a local definition, copying it from the test-fbc.
Then if Enterprise Contract keep asking to update a task, I should be
able to do it here.

Also enabling hermetic build for this pipeline.

Signed-off-by: Julien Ropé <jrope@redhat.com>