Add Intel's related images to manager.yaml
This is the only place where we should change the RELATED_IMAGES. After we
change them here, we should update the rest with "make bundle" and then commit
the changes created in the CSV file.

In this commit I'm also update the date when it was generated (rounding up a few hours).

For further information, please see https://github.com/openshift/sandboxed-containers-operator/pull/1339
and commit 58a821a5d.

When this commit is merged, all images will be synced accordingly.

Signed-off-by: Daniel Kreling <dkreling@redhat.com>