Add a script that helps deal with Mintmaker PRs
We get a lot of dependency updates from Mintmaker.
Reviewing/merging them one by one is taking a lot of time.

This script is meant to help dealing with those pull request.
It will:
- create a local git branch
- identify all the pull requests from Mintmaker
- cherry-pick them one by one to the local branch

The script is ignoring any commit that conflicts.
It also ignores PRs marked as "do not merge", and nudge PRs from Konflux
(that should be worked on separately).

The script is meant to work on all our repositories.
For that purpose, the default values used for branch/remote/repo can be
overridden by environment variables, or using command line parameters.

The resulting branch should be reviewed and pushed as a single PR.
The script doesn't create the PR, and doesn't close the related PRs from Mintmaker.

Signed-off-by: Julien Ropé <jrope@redhat.com>