ibmcloud: check worker node version
Add a function to check the cloud provider in the DaemonSet install
script. If the provider is ibmcloud, retrieve the node's version and
use it to get the extension image for that specific version instead of
using the environment variable. This fixes issues where worker nodes
run a different OCP version than the control plane.

Signed-off-by: Patrik Fodor <patrik.fodor@ibm.com>