test-fbc: allow changelog.sh to log image additions properly
When a new image is added, there is nothing to compare to, and the script fails.
This commit catches this situation and displays the list of new images, if there
is more than one.

Note that we still assume there is only one image *update* per commit in the
general case - if we find more than one change, we currently only log the changes
for the first one.
This is consistent with Konflux's current behavior of making one commit per
image change.

Signed-off-by: Julien Ropé <jrope@redhat.com>