test-fbc: do not try to log changes for version string updates
When we update the version strings, we generate changes that can't be compared
to previous version in git.
Since we use comment markers for the strings to update, the marker can be used
in the script to catch this situation.
This commit makes sure we don't try to do an image changelog for version bumps.

Signed-off-by: Julien Ropé <jrope@redhat.com>