test-fbc: make it possible to call changelog.sh with arbirary commit ranges
It is cumbersome to verify what commits are changing the catalog-template.yaml
file. The test-fbc build is fine, as it is, by definition, running on such a
change. But when we invoke the script locally, we may want more flexibility.

This commit uses the range provided as parameter, and gets the first and last
commit in this range that impacts catalog-template.yaml.

Signed-off-by: Julien Ropé <jrope@redhat.com>